+++ public/index.html
@@ -21,7 +21,9 @@
   <body>
     <div id="root"></div>
   </body>
+  <!-- ELECTRON WEBPACK DEV SERVER -->
   <!-- <script src="http://localhost:4172/bundle.js"></script> -->
+  <!-- ELECTRON PRODUCTION BUNDLE -->
   <!-- <script src="../bundle.js"></script> -->
 </html>
