+++ src/components/home/LightItem.js
@@ -1,5 +1,5 @@
 import React, { useRef } from "react";
-import { convertHSBToColor, compatibleText } from "../utils";
+import { convertHSBToColor, compatibleText } from "../../utils";
 
 const LightItem = ({ light, toggle, alert, active, onSelect, onDim }) => {
   const color = convertHSBToColor(light.state);
