+++ src/components/home/LightsList.js
@@ -10,9 +10,9 @@ import {
   alertRoom,
   toggleRoom,
   setActiveRoom
-} from "../actions";
+} from "../../actions";
 import LightItem from "./LightItem";
-import RoomItem from "./RoomItem";
+import RoomItem from "../RoomItem";
 
 const LightsList = ({
   lights,
